+++ src/e2e/angular/project/gulpfile.js
@@ -1,54 +1,36 @@
-// Sample gulpfile
+"use strict";
+
 const gulp = require("gulp");
 const typescript = require("typescript");
 const buildTools = require("./local-web-build-tools");
 
-const projectOptions = buildTools.config.DEFAULT_PROJECT_OPTIONS;
-projectOptions.root = __dirname;
-
-buildTools.projectTasks.registerAll(gulp, {
-  project: projectOptions,
-  tslintOptions: {},
-  install: {}
-});
-
-
-const serverTarget = {
-  type: "node",
-  baseDir: "server",
-  scripts: ["server/**/*.ts", "app/**/*.ts"],
-  typeRoots: ["./custom-typings", "../typings/globals", "../typings/modules", "../node_modules/@types"],
-  mainModule: "server/main",
-};
-
-buildTools.targetGenerators.node.generateTarget(
-  gulp,
-  "server",
+// Project-wide options
+const projectOptions = Object.assign(
+  {},
+  buildTools.config.DEFAULT_PROJECT_OPTIONS,
   {
-    project: projectOptions,
-    target: serverTarget,
-    tsOptions: {
+    root: __dirname
+  }
+);
+
+// Angular universal server target
+// This will look better with ES7 object spread:
+// const serverTarget = {...buildTools.config.ANGULAR_SERVER_TARGET, name: "foo", scripts: ["bar/*.ts"]}
+const serverTarget = Object.assign(
+  {},
+  buildTools.config.ANGULAR_SERVER_TARGET,
+  {
+    typescriptOptions: {
       skipLibCheck: true,
       typescript: typescript,
       lib: ["es6", "dom"]
     },
-    pug: [
-      {
-        // Name of the gulp sub-task
-        name: "app",
-        // Files to process
-        files: ["app/**/*.pug"],
-        // gulp-pug options
+    pug: buildTools.config.ANGULAR_SERVER_TARGET.pug
+      .map(config => Object.assign({}, config, {
         options: {locals: {locale: "en_US"}}
-      }
-    ],
-    sass: [
-      {
-        // Name of the gulp sub-task
-        name: "app",
-        // Files to process
-        files: ["app/**/*.scss"]
-      }
-    ]
+      }))
   }
 );
+
+buildTools.projectTasks.registerAll(gulp, projectOptions);
+buildTools.targetGenerators.node.generateTarget(gulp, projectOptions, serverTarget);