+++ src/e2e/node-lib/project/gulpfile.js
@@ -1,12 +1,28 @@
 // Sample gulpfile
 const gulp = require("gulp");
+const typescript = require("typescript");
 const buildTools = require("./local-web-build-tools");
 
-const projectOptions = buildTools.config.DEFAULT_PROJECT_OPTIONS;
-projectOptions.root = __dirname;
+// Project-wide options
+const projectOptions = Object.assign(
+  {},
+  buildTools.config.DEFAULT_PROJECT_OPTIONS,
+  {
+    root: __dirname
+  }
+);
 
-buildTools.projectTasks.registerAll(gulp, {
-  project: projectOptions,
-  tslintOptions: {},
-  install: {}
-});
+// Node library
+const libTarget = Object.assign(
+  {},
+  buildTools.config.LIB_TARGET,
+  {
+    typescriptOptions: {
+      skipLibCheck: true,
+      typescript: typescript
+    }
+  }
+);
+
+buildTools.projectTasks.registerAll(gulp, projectOptions);
+buildTools.targetGenerators.node.generateTarget(gulp, projectOptions, libTarget);